move to pointwise_ops
