+++ spmd/tensor/ops/pointwise_ops.py
@@ -1,6 +1,11 @@
 # Copyright (c) Meta Platforms, Inc. and affiliates
+from typing import cast
+
 from spmd.tensor.api import DTensor
 from spmd.tensor.ops.common_rules import linear_pointwise_rule, pointwise_rule
+from spmd.tensor.placement_types import DTensorSpec, Replicate, _Partial
+from spmd.tensor.dispatch import OpSchema, OutputSharding
+from spmd.tensor.ops.utils import register_prop_rule
 
 # leave the remaining pointwise_ops list here for convenience,
 # Below ops are some pointwise ops that are yet to be supported,
@@ -362,3 +367,26 @@ for op in linear_pointwise_ops:
 
 for op in pointwise_ops:
     DTensor._op_to_rules[op] = pointwise_rule
+
+
+@register_prop_rule("aten.native_dropout.default")
+def dropout_rule(op_schema: OpSchema) -> OutputSharding:
+    self_spec = cast(DTensorSpec, op_schema.args_schema[0])
+
+    # TODO: We are specializing dropout_rule now because it's
+    # a non-deterministic algorithm, and replication does not
+    # not support non-deterministic op yet. We should remove
+    # this rule and make dropout to use pointwise rule instead
+    # once we support non-deterministic op.
+    replicate_or_partial = False
+    for placement in self_spec.placements:
+        if isinstance(placement, (Replicate, _Partial)):
+            replicate_or_partial = True
+            break
+
+    if replicate_or_partial:
+        return OutputSharding(
+            None, failed_reason="Dropout with replication is not supported yet!"
+        )
+    else:
+        return OutputSharding(self_spec)
